fix(ios): use NSInteger for setFollowingPerspective parameter type
The perspective parameter was declared as `double` but immediately cast
to `NSInteger`. Using `NSInteger` directly avoids a lossy implicit
floating-point-to-integer conversion and matches the expected enum type.

+++ ios/react-native-navigation-sdk/NavViewModule.mm
@@ -476,13 +476,13 @@ static NavViewModule *sharedInstance = nil;
 }
 
 - (void)setFollowingPerspective:(NSString *)nativeID
-                    perspective:(double)perspective
+                    perspective:(NSInteger)perspective
                         resolve:(RCTPromiseResolveBlock)resolve
                          reject:(RCTPromiseRejectBlock)reject {
   NavViewController *viewController = [self getViewControllerForNativeID:nativeID];
   if (viewController) {
     dispatch_async(dispatch_get_main_queue(), ^{
-      [viewController setFollowingPerspective:@((NSInteger)perspective)];
+      [viewController setFollowingPerspective:@(perspective)];
       resolve(nil);
     });
   } else {
